requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
dj-database-url==2.0.0
Django==4.1.7
django-bootstrap-form==3.4
django-crispy-forms==2.0
django-environ==0.9.0
django-on-heroku==1.1.2
django-recaptcha==3.0.0
django-tempus-dominus==5.1.2.17
django-tinymce==3.5.0
django-widget-tweaks==1.4.12
fontawesomefree==6.3.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
pysqlite3-binary==0.5.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
whitenoise==6.4.0
